+++ packages/runtime/src/execute/expression.ts
@@ -9,7 +9,7 @@ import clone from '../util/clone';
 import {
   InputError,
   TimeoutError,
-  UserError,
+  JobError,
   assertAdaptorError,
   assertImportError,
   assertRuntimeCrash,
@@ -17,6 +17,11 @@ import {
   assertSecurityKill,
 } from '../errors';
 import { NOTIFY_JOB_COMPLETE, NOTIFY_JOB_START } from '../events';
+
+export type ExecutionErrorWrapper = {
+  state: any;
+  error: any;
+};
 
 export default (
   ctx: ExecutionContext,
@@ -77,6 +82,8 @@ export default (
       // return the final state
       resolve(finalState);
     } catch (e: any) {
+      // whatever initial state looks like now, clean it and report it back
+      const finalState = prepareFinalState(opts, initialState);
       duration = Date.now() - duration;
       let finalError;
       try {
@@ -85,12 +92,12 @@ export default (
         assertRuntimeCrash(e);
         assertSecurityKill(e);
         assertAdaptorError(e);
-        finalError = new UserError(e);
+        finalError = new JobError(e);
       } catch (e) {
         finalError = e;
       }
 
-      reject(finalError);
+      reject({ state: finalState, error: finalError } as ExecutionErrorWrapper);
     }
   });