+++ packages/runtime/test/runtime.test.ts
@@ -383,7 +383,7 @@ test('log errors, write to state, and continue', async (t) => {
 
   t.truthy(result.errors);
   t.is(result.errors.a.message, 'test');
-  t.is(result.errors.a.type, 'UserError');
+  t.is(result.errors.a.type, 'JobError');
 
   t.truthy(logger._find('error', /failed job a/i));
 });