+++ packages/engine-multi/CHANGELOG.md
@@ -1,5 +1,13 @@
 # engine-multi
 
+## 0.1.9
+
+### Patch Changes
+
+- Updated dependencies [419f276]
+- Updated dependencies [0e66f5a]
+  - @openfn/runtime@0.1.2
+
 ## 0.1.8
 
 ### Patch Changes